Moved test to dedicated test class
